requirements.txt
asgiref==3.10.0
astroid==4.0.2
colorama==0.4.6
dill==0.4.0
Django==5.2.7
django-cleanup==9.0.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
isort==7.0.0
mccabe==0.7.0
mysqlclient==2.2.7
pillow==12.0.0
platformdirs==4.5.0
PyJWT==2.10.1
pylint==4.0.2
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.3
tzdata==2025.2
